Move ImportBlock into consensus-common
