+++ core/network/src/chain.rs
@@ -20,7 +20,8 @@ use client::{self, Client as SubstrateClient, ClientInfo, BlockStatus, CallExecu
 use client::error::Error;
 use consensus::BlockImport;
 use runtime_primitives::traits::{Block as BlockT, Header as HeaderT, NumberFor};
-use runtime_primitives::generic::{BlockId, ImportBlock, ImportResult};
+use runtime_primitives::generic::{BlockId};
+use consensus::{ImportBlock, ImportResult};
 use runtime_primitives::Justification;
 use primitives::{Blake2Hasher, AuthorityId};
