+++ core/service/Cargo.toml
@@ -20,6 +20,7 @@ substrate-keystore = { path = "../../core/keystore" }
 sr-io = { path = "../../core/sr-io" }
 sr-primitives = { path = "../../core/sr-primitives" }
 substrate-primitives = { path = "../../core/primitives" }
+substrate-consensus-common = { path = "../../core/consensus/common" }
 substrate-network = { path = "../../core/network" }
 substrate-client = { path = "../../core/client" }
 substrate-client-db = { path = "../../core/client/db" }
